+++ lib/src/models.dart
@@ -5,7 +5,6 @@ export 'models/api/channel_min.dart';
 export 'models/api/video.dart';
 export 'models/api/video_full.dart';
 export 'models/api/comment.dart';
-export 'models/api/video_metadata.dart';
 export 'models/api/song.dart';
 export 'models/api/video_live_info.dart';
 export 'models/paginated_result.dart';