+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - **Breaking**: Changed `HolodexClient::getVTuberCollabs` parameters to an optional `ChannelVideoFilter`
 - **Breaking**: Removed `VideoWithCommentsList`
 - **Breaking**: Removed `VideoWithComments`, use `VideoFull` instead
+- **Breaking**: Removed `VideoMetadata`, use `VideoFull` instead
 - **Breaking**: Renamed `client.dart` to `holodex_client.dart`. Technically a breaking change, but only if
 implementation was imported.
 - **Breaking**: Replace usage of package `EnumToString` with Dart 2.17 enhanced enums.